+++ PHY480_HW2/integ_hw2_routines.h
@@ -1,6 +1,6 @@
-extern double simpsons_rule ( int num_pts, float x_min, float x_max,
-                       float (*integrand) (float x) ); //Simpsons Rule
-extern double milne_rule ( int num_pts, float x_min, float x_max,
-                      float (*integrand) (float x) );//Milne Rule
+extern double simpsons_rule ( int num_pts, double x_min, double x_max,
+                       double (*integrand) (double x) ); //Simpsons Rule
+extern double milne_rule ( int num_pts, double x_min, double x_max,
+                      double (*integrand) (double x) );//Milne Rule
 extern double my_gsl_rule(int i); //Gauss-Kronrod 21 point
-//end: function prototypes 
+//end: function prototypes
